common.requirements.txt
aiosqlite~=0.20.0
alembic~=1.13.1
anyio~=3.7.1
apprise~=1.8.0
asyncpg~=0.29.0
attrs~=23.2.0
beanie~=1.26.0
cachetools~=5.3.3
certifi~=2024.2.2
cffi~=1.16.0
click~=8.1.7
cloudpickle~=3.0.0
colorama~=0.4.6
coolname~=2.2.0
croniter~=2.0.5
cryptography~=42.0.7
con
dateparser~=1.2.0
distro~=1.9.0
dnspython~=2.6.1
docker~=6.1.3
email_validator~=2.1.1
fastapi~=0.111.0
fsspec~=2024.5.0
graphviz~=0.20.3
greenlet~=3.0.3
griffe~=0.45.2
h11~=0.14.0
h2~=4.1.0
hpack~=4.0.0
httpcore~=1.0.5
httptools~=0.6.1
httpx~=0.27.0
humanize~=4.9.0
hyperframe~=6.0.1
idna~=3.7
importlib_resources~=6.1.3
iniconfig~=2.0.0
itsdangerous~=2.2.0
Jinja2~=3.1.4
jsonpatch~=1.33
jsonpointer~=2.4
jsonschema~=4.22.0
kubernetes~=29.0.0
Mako~=1.3.5
Markdown~=3.6
MarkupSafe~=2.1.5
mdurl~=0.1.2
motor~=3.4.0
oauthlib~=3.2.2
openai~=1.30.4
orjson~=3.10.3
packaging~=24.0
pathspec~=0.12.1
pendulum~=3.0.0
pluggy~=1.5.0
prefect~=2.19.2
pyasn1~=0.6.0
pyasn1_modules~=0.4.0
pycparser~=2.22
pydantic~=2.7.2
pydantic_core~=2.18.3
pydantic_settings~=2.2.1
Pygments~=2.18.0
pymongo~=4.7.2
pytest~=8.2.1
python-dateutil~=2.9.0.post0
python-dotenv~=0.21.0
python-multipart~=0.0.9
python-slugify~=8.0.4
pytz~=2024.1
PyYAML~=6.0.1
readchar~=4.1.0
referencing~=0.35.1
requests~=2.32.2
rich~=13.7.1
rsa~=4.9
shellingham~=1.5.4
six~=1.16.0
sniffio~=1.3.1
SQLAlchemy~=2.0.30
starlette~=0.37.2
toml~=0.10.2
typer~=0.12.3
typing_extensions~=4.12.0
tzlocal~=5.2
ujson~=5.10.0
urllib3~=2.2.1
uvicorn~=0.28.1
uvloop~=0.19.0
watchfiles~=0.21.0
websocket-client~=1.8.0
websockets~=12.0

confluent-kafka~=2.4.0
kafka-python~=2.0.2

pip~=24.0